backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
python-dotenv
pydantic-settings
pydantic>=2.0
httpx>=0.27
